+++ config/app.php
@@ -1,0 +1,17 @@
+<?php
+
+return [
+    'providers' => [
+        'defaults' => true,
+        'disabled' => [],
+        'extra' => [],
+        'discover' => true,
+        'path' => base_path('bootstrap/providers'),
+        'packages' => [
+            'discover' => true,
+            'dont_discover' => [],
+            'cache' => true,
+            'cache_path' => base_path('bootstrap/cache/packages.php'),
+        ],
+    ],
+];